Refactor: Move previewer classes to a separate module
- Created a new `preview.py` module to house the `HtmlPreview` and `JSONPreview` classes.
- Moved the JSON rendering logic from `renderer.py` into the `JSONPreview` class.
- Updated `renderer.py` to import the previewer classes from the new module and use the refactored implementation.
- This change improves code organization and separation of concerns.
